# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Categorías y Productos",
        "description": "Listado de URI'S de la API Categorías y Productos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Listado de los registros de todas las categorías",
                "operationId": "66bfb3e0bf7747700be142ec5329dfae",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Ropa"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Todo tipo de ropas"
                                            },
                                            "categoria_id": {
                                                "description": "ID de la categoría a la que pertenece el producto",
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-08-23T00:09:16.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-08-23T12:33:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crea una nueva categoría",
                "description": "Crear una nueva categoría",
                "operationId": "acace4e67a5f9bbaa41b456c95b2bd05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Categoria": {
                                        "required": [
                                            "nombre"
                                        ],
                                        "properties": {
                                            "nombre": {
                                                "description": "Nombre de la categoría",
                                                "type": "string",
                                                "example": "Electrónica"
                                            },
                                            "descripcion": {
                                                "description": "Descripción de la categoría",
                                                "type": "string",
                                                "example": "Artículos relacionados con electrónica"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria creada con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "Categoria.nombre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo nombre es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se pudo crear la categoria."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}